+++ lib/plausible/stats/query_result.ex
@@ -70,7 +70,10 @@ defmodule Plausible.Stats.QueryResult do
 
   defp meta(query) do
     %{
-      warning:
+      imports_included: if(query.include.imports, do: query.include_imported, else: nil),
+      imports_skip_reason:
+        if(query.skip_imported_reason, do: Atom.to_string(query.skip_imported_reason), else: nil),
+      imports_warning:
         case query.skip_imported_reason do
           :unsupported_query -> @imports_unsupported_query_warning
           _ -> nil